utils/password.go
package utils

import (
	"fmt"

	"golang.org/x/crypto/bcrypt"
)

//function to hash password
func HashPassword(password string) (string, error) {
	pass, err := bcrypt.GenerateFromPassword([]byte(password), bcrypt.DefaultCost)
	if err != nil {
		return "", fmt.Errorf("failed to hash password: %w", err)
	}
	return string(pass), nil
}

//function to compare the entered password with the db hash
func CheckPassword(password string, hashedPassword string) error {
	return bcrypt.CompareHashAndPassword([]byte(hashedPassword), []byte(password))
}
